+++ src/bindings/matrix.rs
@@ -18,6 +18,8 @@ pub trait Matrix:
     + MulAssign<Self>
     + Mul<Self::Scalar, Output = Self>
     + MulAssign<Self::Scalar>
+    + Div<Self::Scalar, Output = Self>
+    + DivAssign<Self::Scalar>
     + Neg<Output = Self>
     + AbsDiffEq<Epsilon = <Self::Scalar as AbsDiffEq>::Epsilon>
 {
@@ -36,6 +38,8 @@ pub trait Matrix:
     fn inverse(&self) -> Self;
     #[must_use]
     fn determinant(&self) -> Self::Scalar;
+    #[must_use]
+    fn abs(&self) -> Self;
 }
 
 /// Primitive 2x2 matrix.
@@ -203,6 +207,7 @@ macro_rules! impl_matrix {
             forward_impl!($glam_ty => fn determinant(&self) -> $scalar);
             forward_impl!($glam_ty => fn transpose(&self) -> Self);
             forward_impl!($glam_ty => fn inverse(&self) -> Self);
+            forward_impl!($glam_ty => fn abs(&self) -> Self);
         }
     };
 }