+++ src/point.rs
@@ -13,7 +13,7 @@ use bytemuck::{Pod, TransparentWrapper, Zeroable};
 
 use crate::{
     bindings::prelude::*,
-    scalar::{FloatScalar, SignedScalar},
+    scalar::{FloatScalar, IntScalar, SignedScalar},
     AsRaw, FromRaw, Scalar, ToRaw, Unit, Vector2, Vector3, Vector4,
 };
 use core::ops::Mul;
@@ -130,6 +130,34 @@ macro_rules! float_point_interface {
             pub fn distance(self, other: Self) -> T::Scalar;
             #[doc = "Distance squared"]
             pub fn distance_squared(self, other: Self) -> T::Scalar;
+            #[doc = "Midpoint between two points"]
+            pub fn midpoint(self, rhs: Self) -> Self;
+            #[doc = "Moves towards rhs based on the value d."]
+            pub fn move_towards(self, rhs: Self, d: T::Scalar) -> Self;
+        );
+    };
+}
+
+macro_rules! int_point_interface {
+    ($vector:ty, $see_also_doc_ty:ty) => {
+        crate::forward_to_raw!(
+            $see_also_doc_ty =>
+            #[doc = "Returns a vector containing the saturating addition of self and rhs."]
+            pub fn saturating_add(self, rhs: $vector) -> Self;
+            #[doc = "Returns a vector containing the saturating subtraction of self and rhs."]
+            pub fn saturating_sub(self, rhs: $vector) -> Self;
+            #[doc = "Returns a vector containing the saturating multiplication of self and rhs."]
+            pub fn saturating_mul(self, rhs: $vector) -> Self;
+            #[doc = "Returns a vector containing the saturating division of self and rhs."]
+            pub fn saturating_div(self, rhs: $vector) -> Self;
+            #[doc = "Returns a vector containing the wrapping addition of self and rhs."]
+            pub fn wrapping_add(self, rhs: $vector) -> Self;
+            #[doc = "Returns a vector containing the wrapping subtraction of self and rhs."]
+            pub fn wrapping_sub(self, rhs: $vector) -> Self;
+            #[doc = "Returns a vector containing the wrapping multiplication of self and rhs."]
+            pub fn wrapping_mul(self, rhs: $vector) -> Self;
+            #[doc = "Returns a vector containing the wrapping division of self and rhs."]
+            pub fn wrapping_div(self, rhs: $vector) -> Self;
         );
     };
 }
@@ -313,6 +341,10 @@ impl<T: Unit> Point2<T> {
         glam::Vec2 =>
         #[doc = "Extend with z-component to [`Point3`]."]
         pub fn extend(self, z: T::Scalar) -> Point3<T>;
+        #[doc = "Replace the x-component with a new value."]
+        pub fn with_x(self, x: T::Scalar) -> Self;
+        #[doc = "Replace the y-component with a new value."]
+        pub fn with_y(self, y: T::Scalar) -> Self;
     );
 
     point_interface!(Point2, Vector2);
@@ -341,6 +373,14 @@ where
 
     crate::forward_float_ops!(glam::BVec2, glam::Vec2);
     float_point_interface!(glam::Vec2);
+}
+
+impl<T> Point2<T>
+where
+    T: Unit,
+    T::Scalar: IntScalar,
+{
+    int_point_interface!(Vector2<T>, glam::IVec2);
 }
 
 impl<T: Unit> Point3<T> {
@@ -383,6 +423,12 @@ impl<T: Unit> Point3<T> {
         pub fn extend(self, w: T::Scalar) -> Point4<T>;
         #[doc = "Truncate to [`Point2`]."]
         pub fn truncate(self) -> Point2<T>;
+        #[doc = "Replace the x-component with a new value."]
+        pub fn with_x(self, x: T::Scalar) -> Self;
+        #[doc = "Replace the y-component with a new value."]
+        pub fn with_y(self, y: T::Scalar) -> Self;
+        #[doc = "Replace the z-component with a new value."]
+        pub fn with_z(self, z: T::Scalar) -> Self;
     );
 
     point_interface!(Point3, Vector3);
@@ -414,6 +460,14 @@ where
 
     crate::forward_float_ops!(glam::BVec3, glam::Vec3);
     float_point_interface!(glam::Vec3);
+}
+
+impl<T> Point3<T>
+where
+    T: Unit,
+    T::Scalar: IntScalar,
+{
+    int_point_interface!(Vector3<T>, glam::IVec3);
 }
 
 impl<T: Unit> Point4<T> {
@@ -458,6 +512,14 @@ impl<T: Unit> Point4<T> {
         glam::Vec4 =>
         #[doc = "Truncate to [`Point3`]."]
         pub fn truncate(self) -> Point3<T>;
+        #[doc = "Replace the x-component with a new value."]
+        pub fn with_x(self, x: T::Scalar) -> Self;
+        #[doc = "Replace the y-component with a new value."]
+        pub fn with_y(self, y: T::Scalar) -> Self;
+        #[doc = "Replace the z-component with a new value."]
+        pub fn with_z(self, z: T::Scalar) -> Self;
+        #[doc = "Replace the w-component with a new value."]
+        pub fn with_w(self, w: T::Scalar) -> Self;
     );
 
     point_interface!(Point4, Vector4);
@@ -492,6 +554,14 @@ where
 
     crate::forward_float_ops!(glam::BVec4, glam::Vec4);
     float_point_interface!(glam::Vec4);
+}
+
+impl<T> Point4<T>
+where
+    T: Unit,
+    T::Scalar: IntScalar,
+{
+    int_point_interface!(Vector4<T>, glam::IVec4);
 }
 
 impl<T> Point3<T>