+++ src/scalar.rs
@@ -16,6 +16,7 @@ pub trait Scalar:
     + num_traits::NumOps
     + num_traits::NumCast
     + approx::AbsDiffEq<Epsilon = Self>
+    + From<bool>
 {
     /// The underlying 2D vector type for this scalar ([`glam::Vec2`],
     /// [`glam::DVec2`], [`glam::IVec2`], or [`glam::UVec2`]).
@@ -109,6 +110,16 @@ pub trait FloatScalar:
     /// True if the number is not NaN and not infinity.
     #[must_use]
     fn is_finite(self) -> bool;
+}
+
+/// Integer scalar types.
+pub trait IntScalar: Scalar<Vec2 = Self::Vec2i, Vec3 = Self::Vec3i, Vec4 = Self::Vec4i> {
+    #[doc(hidden)]
+    type Vec2i: bindings::Vector2<Scalar = Self> + bindings::IntegerVector;
+    #[doc(hidden)]
+    type Vec3i: bindings::Vector3<Scalar = Self> + bindings::IntegerVector;
+    #[doc(hidden)]
+    type Vec4i: bindings::Vector4<Scalar = Self> + bindings::IntegerVector;
 }
 
 impl Scalar for f32 {